+++ webman-amplifier/content-shortcode-posts-wm_staff.php
@@ -13,7 +13,8 @@
  * @package     WebMan Amplifier
  * @subpackage  Shortcodes
  *
- * @version  1.6
+ * @since    1.0
+ * @version  1.9.1
  *
  * @uses        array $helper  Contains shortcode $atts array plus additional helper variables.
  */
@@ -27,10 +28,10 @@ if ( $helper['link'] ) {
 }
 ?>
 
-<article class="<?php echo $helper['item_class']; ?>"<?php echo wm_schema_org( 'person' ); ?>>
+<article class="<?php echo esc_attr( $helper['item_class'] ); ?>"<?php echo wm_schema_org( 'person' ); ?>>
 
 	<div class="wm-posts-element wm-html-element title"><?php
-		echo '<' . $helper['atts']['heading_tag'] . wm_schema_org( 'name' ) . '>';
+		echo '<' . tag_escape( $helper['atts']['heading_tag'] ) . wm_schema_org( 'name' ) . '>';
 
 			echo $link_output[0];
 
@@ -38,7 +39,7 @@ if ( $helper['link'] ) {
 
 			echo $link_output[1];
 
-		echo '</' . $helper['atts']['heading_tag'] . '>';
+		echo '</' . tag_escape( $helper['atts']['heading_tag'] ) . '>';
 	?></div>
 
 	<?php